requirements.txt
ultralytics==8.2.103
opencv-python==4.10.0.84
numpy==1.26.4
PyYAML==6.0.2
